ci: Stop Docker cache explosion
We _fixed_ this yesterday, and immediately had it again. We had a
cache of ~143GB! We also don't have cache cleanup.

**CI**

Instead, only cache _to_ `main` when running CI. On branches/PRs
though, cache _from_ `main`. The builder layer is the one that
installs dependencies, and is the biggest layer. That's the key one
we want to cache, and reuse.

Previously, we didn't have any caching anyway! So this is still an improvement.

**CD**

We were caching in GHA, even though we use ECR. That was an oversight
from me. Instead, now cache _to_ and _from_ our singular ECR (as
opposed to, per environment). A `buildcache` tag is used, as is
convention[^1].

This helps prevent cache explosion in GitHub and ideally helps speed
up building and pushing to ECR, for CD. Along with this, I've added
tags for SHA and `latest`, which can be helpful for lineage, and
convenience, respectively.

**Appendix**

[^1]: https://github.com/search?q=language%3Ayaml%20%2Fcache-from.*buildcache%2F&type=code