test: Schedule staging runs for validation
We want to run the pipeline(s) more frequently, especially whilst we
aren't runnning them frequently in production.

This allows for schedules to include parameters. It's hard-coded for
now. I think in the future we could read the classifier specs., and
filter some out of there.

The approach follows the example from Prefect[^1].

[^1]: https://docs.prefect.io/v3/how-to-guides/deployments/create-schedules#schedule-parameters-in-python