feat: sort spec fields in a less silly way
Not meaning to cast shade on alphabetic ordering, but it doesnt make
sense in this context where we'd like the key identifiers to be at the
top. It turns out that is what pydantic was thinking too in preserving
the order of the defined fields when dumping the model, but the yaml
dump by defult sorts the keys alphabetically. This switches that
behaviour off