feat: integrate classifier specific filtering
This builds on the introduction of a `dont_run_on` field for classifier
specs. Here the logic is added to inference to respond to the filtering
metadata and skip documents of the given sources.

I've included an artifact at the begining to write out the removal details
early, as a substitute for the previous sabin specific filter which logged
details.