feat: conditionally run on gpus
This meant updating the `map_as_sub_flow`. It's a core part of our running
logic, hence there being changes across the flows. The input to the flow
now accepts both a function and paramaters per batch, whilst before the
function had to be set for all batches. This allows inference to
conditionally set to use a gpu flow in a batch if the classifier needs it.