fix: flakey test teardown
The tests used in aggregate have had several issues with read timeouts.
It turns out that this is due to the way we were cleaning up the bucket,
seemingly doing one at a time was resulting in objects not being cleaned
up fast enough.

As well as now doing clean up in bulk, this also sets the objects to
clear up as teardown of their creation, rather than the bucket itself.

Co-authored-by: Kerry Parker <kerry.parker@climatepolicyradar.org>