feat: set coiled 'scheduler' hardware on deploy
It turned out adding a function decorator within the flow code would lead
to problems. Firstly that coiled would spin up two clusters, one from the
config for the workpool itself and another for the function within that.
Secondly, that trying to combine it on the flow itself meant the coiled
decorator would clash with the prefect flow decorator, leading to
problems with typing, async, and validation during deployment.

Setting it in the deployments is a clean approach for the single host per
batch approach. We still spin up a machine per batch so it is distributed
to a degree. Because we are already running in a batch it makes sense for
this first implementation to not be distributed across tasks. We could
follow up on this to run cross multiple workers. Doing so would involve
setting it at the task level rather then the flow level.