feat: update train to support remote gpu training
Adds a flow that can be triggered by the script in order to run the training
in prefect with coiled. For this the script needed to be refactored in
consideration for security and serialisation. The instantiated Console
object cannot be serialised, and the users aws profile & wikibase env
variables can't be used. I wanted to keep the default behaviour as an option
for running training locally, so I set the coiled option to set these up
before triggering the remaining script.