fix: Ensure image has gpu capabilities
Annoyingly this adds 6gb to the image, but its to add gpu support to it
which we need to run on gpus in coiled. I am keen to see what impact this
has on builds in ci, hopefully the GA runner manages and caching will
take some of the edge off but likely we'll need to figure out a strategy
to take the edge off soon.