feat: add json conversion function
It's often useful to generate a json version of the specs, I've found
myself doing this on occasions when wanting this as a parameter for prefect.

I'm resisting adding this as a script, it just seems like overkill. But
the function can be imported in a terminal python session and run when
needed and I think thats sufficient for now.