refactor: Add function to get logger with fallback
`get_logger` returns a standard Python logger[^3].

The intent is to give someone a function that they can call from code
running on Prefect via ECS or Prefect via Coiled or literally anywhere.

> Get a prefect logger. These loggers are intended for internal use within the prefect package.[^4]

Prefect's documentation does say this. A _Prefect logger_ is one that
is named as `prefect`[^5]. I think we still want this, despite the
internal use mention, since in ECS they currently show up as a Prefect
logger, so why not have it show up with the same name, even when not
in a flow or task run context?

TOWARDS PLA-861

[^1]: https://docs.prefect.io/v3/how-to-guides/workflows/add-logging
[^2]: https://docs.prefect.io/v3/api-ref/settings-ref#logging-level
[^3]: https://github.com/PrefectHQ/prefect/blob/main/src/prefect/logging/loggers.py#L64
[^4]: https://docs.prefect.io/v3/api-ref/python/prefect-logging-loggers#get-logger
[^5]: https://github.com/PrefectHQ/prefect/blob/main/src/prefect/logging/loggers.py#L72C21-L72C49