fix: don't train redirects
Downstream in app land, redirects are not currently handled. Letting them
in to the pipeline means we find out at the point of deployment that they
cannot go live, as we saw with Q57/Q58. This changes the get_concept
behaviour to prevent the use of redirects, wether when used directly or
when this script is used by the train or deploy commands.