feat: add audit script for pulling prefect artifacts
Pulls all artifacts within recursive subflows of a given flow. Then groups
the tables by name into single files for follow up analysis.

There have been frequent occasions where we've needed to analyse hundreds
of batch artifacts, and the ui is not remotely an option for this. I was
using a form of this anyway and have been finding it useful. I know others
have had similar ad hoc scripts. So figured it would be benficial to
commit something back to the repo. For this end I have made this one more
generic, so it should work on any flow.