fix(train): Don't implicitly promote in train
Having an AWS environment as an _alias_ means that that version has
been _promoted_. This shouldn't be implicitly done here. It should be
explicitly done in promotion.

The AWS environment is included in the metadata at training.

This means that all models trained since August 2025 have been
implicitly promoted, when they shouldn't have been.

TWOARDS PLA-838