fix: increase connection pool
The connection pool defaults to 10! Meaning that when attempting to run
concurrently there was a significant opportunity for the connections to
time out. This is why the semaphore made a difference to our networking
errors, because it was getting closer to 10. In order to actually get
proper use of the concurrency we can instead increase the connection pool.
Adding a buffer on top seems sensible as it will mitigate some not
releasing back to the pool. I've also increased the default read timeouts
these where directly causing "Response payload is not completed" and
"Read timeout on endpoint" errors, which I believe where indirectly at
fault as they could lead to connections not being released to the pool