feat(inference): Store metadata for runs
We need to know the version of the classifier specification at this
point-in-time. This is for model hot swapping, so we know exactly
which Wikibase, concept, classifier, etc. versions were used.

Whilst aggregation won't load this, we want to start recording it.
Aggregation will be updated after this, to include the necessary data
to make available for indexing.

I did considering storing each classifier spec. in each
Wikibase-Classifier URI in S3, but that would be a huge amount of
writes.

Save some new metadata, for these lineage purposes. I went for the
kitchen sink. Prefect stores flow runs for a finite time. Our
versioning is manual, so it's forgotten 99% of the time. I've added
the commit SHA and branch to the tags, so we'll have that saved. I
thought about using the inferred flow version, that Prefect does, but
decided against that. It should be in the flow run context though.

Writing metadata/config next to results is a common data/ML
engineering approach, so I feel confident in initiating it here.

I've used the same flow run namespacing as seen in aggregation. It
doesn't have the actual inference results in it now, but, it would in
the future, especially when using a nice binary data format.

FIXES PLA-831