feat(index): Use aggregate config in indexing
If the aggregate run's output that's used for this indexing has the
metadata available, enrich the family documents and document passages
with the v2 of concepts.

It was decided in that previous change that this was a less invasive approach.

This now also stores some metadata of the run in S3, for lineage and
durability beyond Prefect.

FIXES PLA-834