feat(aggregate): Store metadata for runs
Similar to #793, do so for aggregation. This data will be loaded by
indexing. Indexing will use that data to enrich the aggregated results.

I considered updating the results that aggregation stored, but found
this approach faster, and keeps the results stable as is. I think we
could refactor them, but outside of the current project.

FIXES PLA-832