build: Cache packages beyond version changes
If we're to properly version our releases, we can't have the packages'
cache being invalidated by changing the version.

Instead, use a binding to install the packages, without relying on
copying the `pyproject.toml` over.

Test:

I tested this by building the image, changing the version, and
rebuilding it. On the rebuild, it didn't redownload the packages.