refactor(inference): Include classifier spec. in artifacts
This data is needed downstream. Introduce this new field and go all
out by including the whole classifier spec.

This will be used in aggregation, to include it in its results. At the
moment, I'm thinking of re-purposing the `name` field of
`Passage.Concept` to stuff it in. Then, in indexing, to conditionally
load it.

However I go about it, I need the data available.

*Test*

I did a flow run[^1]. The results were correctly saved to S3.

TOWARDS PLA-834

[^1]: https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/068f7a56-f5ab-7204-8000-0a4bf4b0ccd8