refactor(aggregate): Include classifier spec. in artifacts
This data is needed downstream.

I thought about saving the v2 concepts, instead of the v1 concepts.
This would've made it hard to create the v1 concepts, out of the v2
concepts, to ensure that both fields are filled. There's downstream
applications that need to be migrated first. I also thought of saving
both v1 and v2, so indexing wouldn't need to do transformations, but
decided against that due to assumed unreasonable increases in object size.

There's an issue in the project to consider deprecation/migration
overall, so that can handle this.

*Test*

I did a flow run[^1]. The results were correctly saved to S3. It
worked when there was a classifier spec., and when there wasn't, so
there's a nice mixed data result.

TOWARDS PLA-834

[^1]: https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/068f7ba0-cc38-79dd-8000-99d75d03a464