fix: Set env var for Wikibase URLs
I saw a failure[^1] for it trying to get the env var, to construct the
property's return value. I think it's a hack that it calls out to the
env for that property, but that can be fixed later.

I've also removed a pointless check that I had.

[^1]: https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/0690a3d2-4aba-7de4-8000-be3b7a8a69b8

FIXES PLA-830