fix: Make all artifacts unique
There aren't constrained runs, that produce these artifacts, and I'm
not sure if Prefect is smart about showing the right version of the
artifact.

This pattern is already used elsewhere.

Since the artifact is no longer overwritten, some of the tests needed
up to be updated to reflect that, as there's some shared state (artifacts).