Aggregate Flow returns State, which wraps AggregateResult
Full Pipeline expects the Aggregate flow to return a State.

If this is successful the result of the Aggregate State is inspected for an
AggregateResult attribute. This attribute wraps any Errors, which are unpacked by the
Full Test flow and logged, whilst pasing the run_output_identifier on to the Indexing flow.

This allows the full pipeline to contine if there are failed documents at the aggregation flow