Bugfix: Full Pipeline completes its run when Aggregation flow fails PLA-1047 (#924)
Allow the full pipeline to complete if there are failures within the `aggregate` flow.

Aggregate flow returns a State, which contains the new Aggregate Result class.

If the State of the Aggregate flow is `Completed` the Aggregate State is inspected for an
AggregateResult attribute. This attribute wraps any errors, which are unpacked by the
Full Pipeline flow and logged, whilst pasing the `run_output_identifier` on to the `Index` flow.

Since the AggregateResult wraps any errors that occur within `aggregate` Flow executions, this
thus allows the Full Pipeline to continue with the next flow, in this case` the `Index` flow.

I've improved the unit tests of `aggregate` and full `pipeline` Flows.

Added slack notifications for failed runs of Full Pipeline flow.

FullPipeline raises ValueError *after* indexing if Aggregate has failures, so it will be marked as failed in Prefect, but lets successful documents be indexed

Closes PLA-1047