refactor(aggregate): Optionally cached results
This carries on with the assumption that we never run with only some
classifier specs. (CS) as opposed to all of what's in the files.

It relies on checking if the CS file has been modified AFTER the
existing aggregation object, if any, exists. That indicates that
there's been a change (add, remove, update), and aggregation needs to
be run again.

This assumes that all existing objects are up-to-date. As of now, I am
doing a big run on everything.

FIXES PLA-1122