refactor(aggregate): Add metadata for aggregate files
The previous approach of relying on timestamps wasn't able to know if
an aggregation was up-to-date.

Instead, now write a "metadata" file that contains the classifier specs.

If it's present, the expected CS are compared to the actual CS, and if
they mismatch, and then aggregation is run on that document.

If there is no metadata file, then aggregation is run.

This means that the number of S3 reads goes to O(n) for n documents vs
O(n * m) for n documents and m classifier specs. O(n) is when every
document is up-to-date. Realistically, this will be O(n - u * m) where
u is a subset of up-to-date documents.

There's a table artifact created with the results, and also a progress
aritfact during reading the metadata objects.

FIXES PLA-1121